requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
celery-redbeat==2.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.1.2
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.31
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
tenacity==8.1.0
tzdata==2022.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
